refactor(femc): rewrite femc driver
